+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 v0.7 includes:
 * Generates local css and javascript files.
+* Added functions to templating:
+  * Capitalizes sentences (removing hyphens).
+  * Getting an anchor id from a member info.
+  * Getting a list of members for type templating.
 
 v0.6 includes:
 * Reads dependency json to download dependency .xml files for use on documentation.